added messages for catch statements, moved userBodyValidation func to util/middleware.ts
